Limit exceptions caught by set_with_retry()
set_with_retry() was disfunctional until
2aad63cce73794430000232e1c3b129c09ff0ac0, as before this commit set()
didn't throw any exceptions, but only returned them.

As set() can throw a multitude of wrapped exceptions and
set_with_retry() only handles _InactiveRpcError() properly (only
exception with a code() method), we need to filter for this exception.
Otherwise, we might just end up with an AttributeError for everything
that is not a _InactiveRpcError.

Using __context__ was a very cool and python3-ish way of getting the
parent exception, but as we now explicitly wrap the exception I opted
for using this (I originally wrote gNMIException without __context__ in
mind and also wanted to make it explicit that we are wrapping another
exception at which the user might want to take a look).